refactor(mainpage): Moved KakaoMap and refactor with zustand store (#9)
* refactor: Moved KakaoMap component and delete useless code

* refactor: Use zustand store instead of props in KakaoMap

* refactor: Remove unused code
